Delete README.md
The Mandala is a Hindu and Buddhist spiritual and ritual symbol that represents the universe. A sacred circle is a Mandala. It represents the womb of creation and mandals are simple geometric shapes constructed from uniform divisions of the circle. The Mandals have long been employed in spiritual practices of pondering on creation and life. The mandala's geometric motifs have layers of meaning. The circle is the basic shape of sacred geometry and the portal between the mundane and spiritual realms.

The Mandala is a geometric design that represents the cosmos. The Mandala is a symbol of self-reconciliation in Jungian psychology. The ‘seed' is the origin of the sacred Mandala. The seed is the center of attraction for external energy. Drawing the energies unfolds the artist's energy. So it represents inner and outside spaces. The Mandala's essence is in the center, and its clutching is around it. So mandala means capturing essence.

References:

github.com/ythecombinator/python-homework/blob/master/turtle/mandala.py

https://www.tutorialspoint.com/turtle-graphics-using-python